feat: add continuous mode
continuous mode: if you need to continue generating faces for the same
dataset using the same configuration, all you need to do is to put the
path of the old dataset (directory where the dataset.csv file is
located), the plugin will load the old config and will disable all the
input fields for the configuration except for the quantity and saving
path.
in case you want to create a new dataset you can change the saving path
and all the input fields will be enabled and their values will be set to
the dafault values.